refactor: move dispatch intersecting to events utils
