+++ frontend/src/lib/utils/events.utils.ts
@@ -1,3 +1,5 @@
+import type { IntersectingDetail } from "$lib/types/intersection.types";
+
 export const emit = <T>({
   message,
   detail,
@@ -10,4 +12,18 @@ export const emit = <T>({
     bubbles: true,
   });
   document.dispatchEvent($event);
+};
+
+export const dispatchIntersecting = ({
+  element,
+  intersecting,
+}: {
+  element: HTMLElement;
+  intersecting: boolean;
+}) => {
+  const $event = new CustomEvent<IntersectingDetail>("nnsIntersecting", {
+    detail: { intersecting },
+    bubbles: false,
+  });
+  element.dispatchEvent($event);
 };
